+++ packages/media-parser/src/has-all-info.ts
@@ -9,9 +9,9 @@ import {hasNumberOfAudioChannels} from './get-number-of-audio-channels';
 import {hasSampleRate} from './get-sample-rate';
 import {getHasTracks} from './get-tracks';
 import {hasVideoCodec} from './get-video-codec';
-import {maySkipVideoData} from './may-skip-video-data/may-skip-video-data';
 import {hasMetadata} from './metadata/get-metadata';
 import type {AllParseMediaFields, Options, ParseMediaFields} from './options';
+import {maySkipVideoData} from './state/may-skip-video-data';
 import type {ParserState} from './state/parser-state';
 
 export const getAvailableInfo = ({
